commonlib/storage: fix int-to-pointer-cast error
When pulling in commonlib/storage/pci_sdhci.c into herobrine, am
seeing an "error: cast to pointer from integer of different size
[-Werror=int-to-pointer-cast]", so fixing that.

BUG=b:254092907
BRANCH=None
TEST=emerge-herobrine coreboot
     Make sure that we can build without errors

Reviewed-on: https://review.coreboot.org/c/coreboot/+/71828
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/commonlib/storage/pci_sdhci.c
@@ -39,7 +39,7 @@ struct sd_mmc_ctrlr *new_mem_sdhci_controller(void *ioaddr)
 
 struct sd_mmc_ctrlr *new_pci_sdhci_controller(pci_devfn_t dev)
 {
-	uint32_t addr;
+	uintptr_t addr;
 
 	addr = pci_s_read_config32(dev, PCI_BASE_ADDRESS_0);
 	if (addr == ((uint32_t)~0)) {
